+++ profiles/preview/preview/appconfiguration/mgmt/appconfiguration/appconfigurationapi/models.go
@@ -8,7 +8,7 @@
 
 package appconfigurationapi
 
-import original "github.com/Azure/azure-sdk-for-go/services/preview/appconfiguration/mgmt/2020-07-01-preview/appconfiguration/appconfigurationapi"
+import original "github.com/Azure/azure-sdk-for-go/services/preview/appconfiguration/mgmt/2021-03-01-preview/appconfiguration/appconfigurationapi"
 
 type ConfigurationStoresClientAPI = original.ConfigurationStoresClientAPI
 type KeyValuesClientAPI = original.KeyValuesClientAPI